Continuous Integration Publish: 71830e3c1b9cd10b52b9095625aefc21f8c7f630 (maint) Remove unnecessary text
While doing the last release, it was noticed that there were extra
lines in the documentation for the choco download command.  This
section should contain only information about the options/switches,
starting with the help switch.

It could be that this line is being added incorrectly via the
GenerateDocs.ps1 script, and we will need to watch for the re-addition
of it when we run that script.